requirements.txt
pyinstaller
customtkinter
playwright
google-api-python-client
google-auth
paho-mqtt
